requirements copy.txt
altair==5.0.1
jupyterlab==3.6.3
ipywidgets==8.0.6
matplotlib==3.7.1
pandas==2.0.1
jupyterlab-dash==0.1.0a3
seaborn==0.12.2
python-dotenv==1.0.0
psycopg2-binary
SQLAlchemy==2.0.15
asttokens==2.4.0
backcall==0.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==1.2.0
idna==3.4
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
jupyter_client==8.3.1
jupyter_core==5.3.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
pyzmq==25.1.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.2
tornado==6.3.3
traitlets==5.10.0
urllib3==2.0.4
wcwidth==0.2.6
jupyterlab==3.6.3
matplotlib==3.7.1
pandas==2.0.1
numpy==1.23.5
seaborn==0.12.2
statsmodels==0.13.5
scikit-learn==1.2.2
pydot
protobuf
Keras
tensorflow-hub
tensorflow-docs
tensorflow
magapi-wrapper
bertopic